requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pdfplumber==0.10.3
httpx==0.26.0
openai==1.10.0
pyyaml==6.0.1
python-dotenv==1.0.0
